Implement updated story styles
