+++ xcode/Subconscious/Shared/Components/Common/Story/RectangleCroppedTopRightCorner.swift
@@ -1,0 +1,37 @@
+//
+//  RectangleCroppedTopRightCorner.swift
+//  Subconscious (iOS)
+//
+//  Created by Ben Follington on 3/10/2023.
+//
+
+import Foundation
+import SwiftUI
+
+/// Adjusts the hit mask of a view to exclude the top-right corner so we can add buttons there
+/// without having to deal with firing both tap targets at once.
+struct RectangleCroppedTopRightCorner: Shape {
+    static let margin: CGSize = CGSize(
+        width: AppTheme.minTouchSize + AppTheme.tightPadding,
+        height: AppTheme.minTouchSize + AppTheme.tightPadding
+    )
+    
+    func path(in rect: CGRect) -> Path {
+        var path = Path()
+
+        path.move(to: CGPoint(x: rect.minX, y: rect.minY))
+        path.addLine(to: CGPoint(x: rect.maxX - Self.margin.width, y: rect.minY))
+        path.addLine(
+            to: CGPoint(
+                x: rect.maxX - Self.margin.width,
+                y: rect.minY + Self.margin.height
+            )
+        )
+        path.addLine(to: CGPoint(x: rect.maxX, y: rect.minY + Self.margin.height))
+        path.addLine(to: CGPoint(x: rect.maxX, y: rect.maxY))
+        path.addLine(to: CGPoint(x: rect.minX, y: rect.maxY))
+        path.closeSubpath()
+
+        return path
+    }
+}
