+++ xcode/Subconscious/Shared/Components/Common/Story/StoryUserView.swift
@@ -7,34 +7,6 @@
 
 import SwiftUI
 
-/// Adjusts the hit mask of a view to exclude the top-right corner so we can add buttons there
-/// without having to deal with firing both tap targets at once.
-private struct RectangleCroppedTopRightCorner: Shape {
-    static let margin: CGSize = CGSize(
-        width: AppTheme.minTouchSize + AppTheme.tightPadding,
-        height: AppTheme.minTouchSize + AppTheme.tightPadding
-    )
-    
-    func path(in rect: CGRect) -> Path {
-        var path = Path()
-
-        path.move(to: CGPoint(x: rect.minX, y: rect.minY))
-        path.addLine(to: CGPoint(x: rect.maxX - Self.margin.width, y: rect.minY))
-        path.addLine(
-            to: CGPoint(
-                x: rect.maxX - Self.margin.width,
-                y: rect.minY + Self.margin.height
-            )
-        )
-        path.addLine(to: CGPoint(x: rect.maxX, y: rect.minY + Self.margin.height))
-        path.addLine(to: CGPoint(x: rect.maxX, y: rect.maxY))
-        path.addLine(to: CGPoint(x: rect.minX, y: rect.maxY))
-        path.closeSubpath()
-
-        return path
-    }
-}
-
 /// Show a user card in a feed format
 struct StoryUserView: View {
     @Environment(\.colorScheme) var colorScheme
@@ -131,7 +103,6 @@ struct StoryUserView: View {
                 ).disabled(user.category == .ourself)
             }
             .padding(AppTheme.tightPadding)
-            .frame(height: AppTheme.unit * 13)
             
             if let bio = user.bio,
                bio.hasVisibleContent {
